// src/app/models/academics/class-assessment-parameter-viewmodel.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ClassAssessmentParameterViewModel
 */
export interface ClassAssessmentParameterViewModel {
    /**
     * Gets or sets ClassId
     */
    classId: Array<number>;
    /**
     * Gets or sets SectionId
     */
    sectionId?: Array<number>;
    /**
     * Gets or sets SchoolBoardId
     */
    schoolBoardId?: number;
    /**
     * Gets or sets AssessmentParameterId
     */
    assessmentParameterId: Array<number>;
    /**
    * get or set
    */
    assesmentCategoryId?: number;
    /**
     * get or set
     */
    assesmentGroupId?: number;
}