// src/app/models/academics/class-assesment-filter-view-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface ClassAssesmentFilterViewModel {
    sortBy: any;
    classes: Array<any>;
    sections: Array<any>;
    schoolBoards: Array<any>;
    assesmentGroups: Array<any>;
    assesmentCategories: Array<any>;
    assesmentSubCategories: Array<any>;
    assesmentParameters: Array<any>;
    sortOrder: 0 | 1 | -1;
    pageNumber: number;
    pageSize: number;
}

/**ValidationMessageView */
export interface ValidationMessageView {
    /**///  */
    // tslint:disable-next-line:no-misused-new
    new(): ValidationMessageView;
    /**Gets or sets list of string mesaages */
    messages: { [key: string]: string };

    /**Gets or sets StatusCode */
    statusCode: number;
}